fix(cards): CardReply & CardReview renamed to CardComment & CardReactiveRating
